requirements.txt
python-nomad==1.1.0